+++ src/handlers/CommandHandler.ts
@@ -1,5 +1,5 @@
 import { Client, Message } from "discord.js";
-import { CommandFactory } from "./CommandFactory";
+import { CommandFactory } from "../cmd/CommandFactory";
 
 export class CommandHandler {
 	factory: CommandFactory;
